+++ cookbooks/chub_oddrest/attributes/default.rb
@@ -1,10 +1,6 @@
 default[:chub_oddrest][:app_url] = "http://artifactory01.nexus.commercehub.com/artifactory/libs-release/com/commercehub/oddservice/%5BRELEASE%5D/oddservice-%5BRELEASE%5D.jar"
 default[:chub_oddrest][:java_heap_size] = "512M"
-# DB
-default[:chub_oddrest][:db][:server] = "SQLVM03"
-default[:chub_oddrest][:db][:name] = "odddev1_odd"
-default[:chub_oddrest][:db][:userName] = "devusr01"
-default[:chub_oddrest][:db][:pass] = "usrdev"
+default[:chub_oddrest][:properties_location] = "/etc/oddrest/orgDataDir.properties"
 
 default[:chub_oddrest][:db][:showSql] = false
 default[:chub_oddrest][:db][:formatSql] = false