+++ cookbooks/chub_sensu/recipes/redis_slave.rb
@@ -1,0 +1,28 @@
+#
+# Cookbook Name:: chub_sensu
+# Recipe:: redis_slave
+#
+# Copyright (C) 2014 CommerceHub
+# 
+# All rights reserved - Do Not Redistribute
+#
+
+master_node = Hash.new
+
+if Chef::Config[:solo]
+  #Define your data as attributes in vagrant
+  master_node = { :address => node.sensu.redis.master.address, :port => node.sensu.redis.master.port }
+else
+  anode = search(:node, "role:sensu_redis_master").first
+  master_node = { :address => anode.fqdn, :port => anode.sensu.redis.master.port }
+end
+
+log "Master Address -> #{master_node[:address]} Port -> #{master_node[:port]} " do
+  level :info    
+end
+
+node.override.redisio.servers = [{:port => node.sensu.redis.port, :slaveof => {:address => master_node[:address], :port => master_node[:port] } }]
+
+include_recipe "redisio"
+include_recipe "redisio::enable"
+include_recipe "chub_sensu::redis_role_xinetd"