+++ cookbooks/chub_sensu_sso/metadata.rb
@@ -4,9 +4,9 @@ maintainer_email 'dcarr@commercehub.com'
 license          'All rights reserved'
 description      'Installs/Configures Sensu in the CommerceHub way for SSO'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.17'
+version          '0.1.18'
 
-depends "sensu"
+depends "sensu", "= 0.8.0"
 depends "base"
 depends "partial_search"