+++ environments/mc_flightcontrol_dev.rb
@@ -1,14 +1,25 @@
 name "mc_flightcontrol_dev"
 description "Mission Control FlightControl Development Environment"
 cookbook_versions({
-	"chub_mc_app" => "= 0.1.0",
+	"chub_mc_app" => "= 0.2.0",
 	"chub_mc_logs" => "= 0.1.0"
 })
 default_attributes( 
 	"chub_mc_app" => {
 		"port" => "9200",
 		"app_name" => "flightcontrol",
-		"jar_file_url" => "http://artifactory01/artifactory/inteng-snapshot/flightcontrol/0.1/flightcontrol.jar"
+		"config_vars" => {
+			"useBridge" => "true",
+			"disableRecompile" => "false",
+			"enableReload" => "false",
+			"mcHost" => "devmissioncontrol",
+			"appHost" => "devmissioncontrol/flightcontrol",
+			"authService" => "devmcauth",
+			"appMachine1" => "dmcflightctrl01",
+			"appMachine2" => "dmcflightctrl02",
+			"pgenServiceHost" => "dmcflightws01"
+		},
+		"jar_file_url" => "http://artifactory01.nexus.commercehub.com/artifactory/inteng-snapshot/com/commercehub/flightcontrol/%5BRELEASE%5D/flightcontrol-%5BRELEASE%5D.jar"
 	},
 	"chub_mc_logs" => {
 		"mc_resource" => "flightcontrol",