+++ environments/mpdev.rb
@@ -1,8 +1,24 @@
 name "mpdev"
 description "dev environment"
 cookbook "base", "= 0.1.16"
-cookbook "chub-buyspace", "= 0.8.0"
-default_attributes "chub-buyspace" => {
+cookbook "chub-klerk", "= 3.0.0"
+cookbook "chub-buyspace", "= 0.10.1"
+default_attributes "chub-klerk" => {
+    "mongo_addresses" => [ "mpqa03:27777", "mpqa05:27777" ],
+    "quartz_database_url" => "jdbc:jtds:sqlserver://sqlvm81/mpqa1_klerk_quartz;user=devusr01;password=usrdev",
+    "hazelcast_group_name" => "mpqa1_klerk",
+    "hazelcast_group_password" => "mpqa1_klerk",
+    "antivirus_hosts" => [ "qa-scanengine.commercehub.com" ],
+    "product_data_event_endpoints" => {
+        "productstream" => "http://mpqa1-api.commercehub.com/klerk/productDataEvents"
+    },
+    "product_data_application_metadata_update_job_finished_event_endpoint" => "jms:queue:KlerkProductDataApplicationMetadataUpdateJobFinishedEventQueue",
+    "dead_product_data_application_metadata_update_job_finished_event_endpoint" => "jms:queue:KlerkProductDataApplicationMetadataUpdateJobFinishedEventDLQ",
+    "sprite_conductor_url" => "http://mpqa05.nexus.commercehub.com:8080/sprite-conductor",
+    "sprite_output_container_prefix" => "mpqa1-",
+    "jms_provider_url" => "jnp://mpqa03.nexus.commercehub.com:7099"
+},
+"chub-buyspace" => {
   "mongo" => {
     "databaseName" => "marketplaceProd_91212bak",
     "replicas" => ['mpqa05.nexus.commercehub.com:27777', 'mpqa03.nexus.commercehub.com:27777'],
@@ -46,10 +62,10 @@ default_attributes "chub-buyspace" => {
     "enabled" => false
   },
   "grails" => {
-    "serverURL" => "http://mpdev1wapp1.nexus.commercehub.com:8080",
-    "apiServerURL" => "http://mpdev1wapp1.nexus.commercehub.com:8080"
+    "serverURL" => "http://mpdev1wapp1.commercehub.com:8080",
+    "apiServerURL" => "http://mpdev1wapp1.commercehub.com:8080"
   },
-  "webServerUrl"=> "http://mpdev1wapp1.nexus.commercehub.com:8080",
+  "webServerUrl"=> "http://mpdev1wapp1.commercehub.com:8080",
   "antivirus" => {
     "hosts" => "['10.10.40.80']",
     "enabled" => true,