+++ environments/odddev1.rb
@@ -39,13 +39,7 @@ default_attributes(
     }
   },
   :chub_oddrest => {
-    :app_url => "http://artifactory01.nexus.commercehub.com/artifactory/libs-release/com/commercehub/oddservice/%5BRELEASE%5D/oddservice-%5BRELEASE%5D.jar",
-    :db => {
-      :server => "SQLVM03",
-      :name => "odddev1_odd",
-      :userName => "devusr01",
-      :pass => "usrdev"
-    }
+    :app_url => "http://artifactory01.nexus.commercehub.com/artifactory/libs-release/com/commercehub/oddservice/%5BRELEASE%5D/oddservice-%5BRELEASE%5D.jar"
   },
   :chub_apartment => {
     :app_url => "http://artifactory01.nexus.commercehub.com/artifactory/libs-release/com/commercehub/apartment/%5BRELEASE%5D/apartment-%5BRELEASE%5D.jar",