+++ roles/netsuitews.rb
@@ -1,0 +1,3 @@
+name "netsuitews"
+description "A linux machine running the Flight WS "
+run_list "recipe[base]", "recipe[chub_mc_webservice::netsuite_webservice]"