+++ cookbooks/chub_yui_combo/CHANGELOG.md
@@ -1,5 +1,14 @@
+# 0.5.1
+- Removed 'sudo' from script to clear the nginx cache
+
+# 0.5.0
+- Added a script to clear the nginx cache
+
+# 0.4.0
+- Added a script to download YUI releases
+
 # 0.3.0
-- Add built-in 'admin' route and 'up.js' file to support monitoring and load-balancer status checks
+- Added built-in 'admin' route and 'up.js' file to support monitoring and load-balancer status checks
 - Fixed issue where routes configured via the routes attribute would cause the combo server to fail to start because thier directories didn't exist yet 
 
 # 0.1.0